requirements.txt
biopython
numpy
pytest
torch
--extra-index-url https://download.pytorch.org/whl/cu121